+++ src/list/ft_parser.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_parser.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nasong <nasong@student.42seoul.kr>         +#+  +:+       +#+        */
+/*   By: sham <sham@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/05 19:53:07 by nasong            #+#    #+#             */
-/*   Updated: 2021/12/05 22:21:37 by nasong           ###   ########.fr       */
+/*   Updated: 2021/12/12 14:10:00 by sham             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "../minishell.h"
 
 int quotes_check(char *str)
 {
@@ -180,6 +180,6 @@ int ft_parser(t_list *cmd_list, char *input)
 	}
 	add_cmd(cmd_list, &word_list);
 	free_str_list(&word_list);
-	//print_cmd_list(cmd_list);
+	print_cmd_list(cmd_list);
 	return (0);
 }
