+++ src/list/ft_string_list.c
@@ -1,30 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_cmd_list.c                                      :+:      :+:    :+:   */
+/*   ft_string_list.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nasong <nasong@student.42seoul.kr>         +#+  +:+       +#+        */
+/*   By: sham <sham@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/12/05 21:57:01 by nasong            #+#    #+#             */
-/*   Updated: 2021/12/05 22:28:03 by nasong           ###   ########.fr       */
+/*   Created: 2021/12/05 21:56:01 by nasong            #+#    #+#             */
+/*   Updated: 2021/12/12 14:05:15 by sham             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "../minishell.h"
 
-void	init_cmd(t_cmd *cmd)
-{
-	cmd->cmd = 0;
-	cmd->arg = 0;
-	init_list(&cmd->arg_list);	
-	init_list(&cmd->discriptor);	
-}
-
-int	free_cmd_list(t_list *list)
+int	free_str_list(t_list *list)
 {
 	t_data *now;
 	t_data *next;
-	t_cmd *target;
 
 	if (list == 0)
 		return (-1);
@@ -34,8 +25,8 @@ int	free_cmd_list(t_list *list)
 	while (now != 0)
 	{
 		next = now->next;
-		target = now->contents;
-		//free
+		free(now->contents);
+		free(now);
 		now = next;
 		list->size--;
 	}
@@ -44,10 +35,10 @@ int	free_cmd_list(t_list *list)
 	return (1);
 }
 
-void print_cmd_list(t_list *list)
+void print_str_list(t_list *list)
 {
 	t_data *now;
-	t_cmd *target;
+	char *target;
 	int index;
 
 	if (list == 0)
@@ -55,15 +46,15 @@ void print_cmd_list(t_list *list)
 		printf("NULL LIST\n");
 		return ;
 	}
+	index = 0;
 	now = list->front;
-	printf("\n=====PRINT CMD LIST / SIZE: %d===\n", list->size);
+	printf("\n---PRINT STR LIST / SIZE: %d---\n", list->size);
 	while (now != 0)
 	{
-		index = 0;
 		target = now->contents;
-		printf("cmd: %s | arg: %s\n", target->cmd, target->arg);
-		print_str_list(&target->arg_list);
+		printf("[%d] %s\n", index, target);
 		now = now->next;
+		index++;
 	}
-	printf("=====PRINT CMD LIST END===\n");
+	printf("---PRINT STR LIST END---\n");
 }
